requirements.txt
# backend/requirements.txt

Flask
Flask-Cors
google-generativeai
newspaper3k